telegram_bots/requirements.txt
python-telegram-bot
python-dotenv
asyncpg
moviepy==1.0.3